+++ packages/typespec-rust/CHANGELOG.md
@@ -5,11 +5,14 @@
 ### Breaking Changes
 
 * Pageable methods will be renamed to start with `list` (e.g. `get_versions` becomes `list_versions`). A warning diagnostic is displayed when such a rename occurs.
+* Sub-clients that specify a `@clientName` decorator will use that client name verbatim instead of having the parent client name as a prefix.
 
 ### Bug Fixes
 
 * Client constructors will now return an error if the endpoint parameter doesn't start with `http[s]`.
 * Added support for unsigned integer types.
+* Preserve `pub(crate)` on sub-client fields that can also be individually initialized.
+* Removed redundant client accessor parameters that can be inherited from the parent client.
 
 ### Other Changes